# cmd/server/main.go
// Command server starts the 6S HTTP API.
package main

import (
	"context"
	"database/sql"
	"fmt"
	"log"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"6s/internal/ai"
	"6s/internal/auth"
	"6s/internal/config"
	"6s/internal/cron"
	"6s/internal/crypto"
	"6s/internal/database"
	"6s/internal/db"
	"6s/internal/i18n"
	"6s/internal/issue"
	"6s/internal/masterdata"
	"6s/internal/notification"
	"6s/internal/observability"
	"6s/internal/report"
	"6s/internal/response"
	"6s/internal/scoring"
	"6s/internal/settings"
	"6s/internal/storage"
)

const readinessVersion = "dev"

var serverStartedAt = time.Now()

type readinessChecks struct {
	Database string `json:"database"`
	Storage  string `json:"storage"`
}

type readinessPayload struct {
	Status        string          `json:"status"`
	DB            string          `json:"db"`
	Version       string          `json:"version"`
	UptimeSeconds int64           `json:"uptime_seconds"`
	Checks        readinessChecks `json:"checks"`
}

func main() {
	cfg, err := config.Load(os.Args[1:])
	if err != nil {
		log.Fatalf("invalid config: %v", err)
	}

	rootCtx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
	defer stop()

	connectCtx, cancelConnect := context.WithTimeout(rootCtx, 30*time.Second)
	defer cancelConnect()
	dbConn, err := database.Connect(connectCtx, cfg.DBDSN, database.DefaultPoolConfig())
	if err != nil {
		if dbConn != nil {
			_ = dbConn.Close()
		}
		log.Fatalf("database connection failed")
	}
	defer func() {
		if err := dbConn.Close(); err != nil {
			log.Printf("error closing db: %v", err)
		}
	}()
	if err := database.ValidateMigrations(connectCtx, dbConn); err != nil {
		log.Fatalf("database schema validation failed: %v", err)
	}

	var cipher *crypto.Cipher
	if cfg.EncryptionKey != "" {
		c, err := crypto.NewCipher(cfg.EncryptionKey)
		if err != nil {
			log.Fatalf("invalid APP_ENCRYPTION_KEY: %v", err)
		}
		cipher = c
	} else if !cfg.DevInsecure {
		log.Fatalf("APP_ENCRYPTION_KEY is required in production")
	}

	r := setupRouter(dbConn, cfg, cipher, nil, rootCtx)

	server := &http.Server{
		Addr:              fmt.Sprintf(":%s", cfg.Port),
		Handler:           r,
		ReadHeaderTimeout: 3 * time.Second,
	}

	serverErrCh := make(chan error, 1)
	go func() {
		if cfg.TLSCert != "" && cfg.TLSKey != "" {
			observability.Log("info", fmt.Sprintf("Server listening with TLS on :%s", cfg.Port), nil)
			serverErrCh <- server.ListenAndServeTLS(cfg.TLSCert, cfg.TLSKey)
		} else {
			observability.Log("info", fmt.Sprintf("Server listening on :%s", cfg.Port), nil)
			serverErrCh <- server.ListenAndServe()
		}
	}()

	select {
	case err := <-serverErrCh:
		if err != nil && err != http.ErrServerClosed {
			log.Fatalf("server terminated: %v", err)
		}
	case <-rootCtx.Done():
		observability.Log("info", "shutdown signal received; draining active connections", nil)
		shutdownCtx, cancelShutdown := context.WithTimeout(context.Background(), 30*time.Second)
		defer cancelShutdown()
		if err := server.Shutdown(shutdownCtx); err != nil {
			observability.Log("error", "server shutdown failed", map[string]any{"error": err.Error()})
		}
	}
}

func setupRouter(dbConn *sql.DB, cfg *config.Config, cipher *crypto.Cipher, ldapClient auth.LDAPClient, parentCtx ...context.Context) *chi.Mux {
	r := chi.NewRouter()

	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(requestLogger)
	r.Use(middleware.Recoverer)
	r.Use(timeoutByRoute)
	r.Use(i18n.Middleware)
	// Liveness is dependency-free so orchestrators restart only stopped processes.
	healthHandler := func(w http.ResponseWriter, _ *http.Request) {
		_ = response.JSON(w, http.StatusOK, map[string]string{
			"status": "ok",
			"db":     "not_checked",
		})
	}
	readinessHandler := func(w http.ResponseWriter, req *http.Request) {
		databaseStatus := "disconnected"
		storageStatus := "unwritable"
		if dbConn != nil {
			ctx, cancel := context.WithTimeout(req.Context(), time.Second)
			err := dbConn.PingContext(ctx)
			cancel()
			if err == nil {
				databaseStatus = "ok"
			}
		}

		storageDir := "./data"
		if cfg != nil && cfg.DataDir != "" {
			storageDir = cfg.DataDir
		}
		if err := observability.CheckStorageWritability(storageDir); err == nil {
			storageStatus = "ok"
		}

		ready := databaseStatus == "ok" && storageStatus == "ok"
		status := "not_ready"
		statusCode := http.StatusServiceUnavailable
		if ready {
			status = "ok"
			statusCode = http.StatusOK
		}
		payload := readinessPayload{
			Status: status, DB: databaseStatus, Version: readinessVersion,
			UptimeSeconds: int64(time.Since(serverStartedAt).Seconds()),
			Checks:        readinessChecks{Database: databaseStatus, Storage: storageStatus},
		}
		if !ready {
			observability.Log("warn", "readiness check failed", map[string]any{
				"database": databaseStatus, "storage": storageStatus,
			})
		}
		_ = response.JSON(w, statusCode, payload)
	}
	r.Get("/api/health", healthHandler)
	r.Head("/api/health", healthHandler)
	r.Get("/api/ready", readinessHandler)
	r.Head("/api/ready", readinessHandler)
	bgCtx := context.Background()
	if len(parentCtx) > 0 && parentCtx[0] != nil {
		bgCtx = parentCtx[0]
	}
	if dbConn != nil && cfg != nil {
		registerAPIRoutes(bgCtx, r, dbConn, cfg, cipher, ldapClient)
	}

	registerStaticRoutes(r)
	return r
}

func registerAPIRoutes(ctx context.Context, r *chi.Mux, dbConn *sql.DB, cfg *config.Config, cipher *crypto.Cipher, ldapClient auth.LDAPClient) {
	queries := db.New(dbConn)
	jwtKey := []byte(cfg.JWTSecret)
	tm := auth.NewTokenManager(jwtKey)

	var trustedProxies []string
	if cfg.TrustedProxies != "" {
		for _, p := range strings.Split(cfg.TrustedProxies, ",") {
			if trimmed := strings.TrimSpace(p); trimmed != "" {
				trustedProxies = append(trustedProxies, trimmed)
			}
		}
	}
	limiter := auth.NewLoginLimiter(trustedProxies)

	ticketMgr := auth.NewTicketManager()
	authHandler := auth.NewHandler(queries, tm, limiter, cipher, ldapClient, ticketMgr)
	authMw := auth.NewMiddleware(tm, queries, ticketMgr)
	adHandler := auth.NewADConfigHandler(queries, cipher, ldapClient)
	permissionHandler := auth.NewPermissionHandler(queries)
	userAdminHandler := auth.NewAdminHandler(queries)
	locationMembershipHandler := auth.NewLocationMembershipHandler(queries)
	teamLocationHandler := auth.NewTeamLocationHandler(queries)
	settingsHandler := settings.NewHandler(queries)
	// Public auth routes
	r.Route("/api/auth", func(ar chi.Router) {
		ar.Post("/login", authHandler.Login)
		ar.Post("/refresh", authHandler.Refresh)
		ar.Get("/setup-status", authHandler.SetupStatus)
		ar.Post("/setup", authHandler.SetupSuperadmin)

		// Authenticated auth routes
		ar.Group(func(pr chi.Router) {
			pr.Use(authMw.Authenticate)
			pr.Post("/ticket", authHandler.CreateTicket)
			pr.Post("/revoke", authHandler.Revoke)
			pr.Get("/sessions", authHandler.Sessions)
		})
	})

	// Config routes (Admin only)
	r.Route("/api/config", func(cr chi.Router) {
		cr.Use(authMw.Authenticate)
		cr.Use(auth.RequirePermission(auth.PermissionADManage))

		cr.Get("/ad", adHandler.GetADConfig)
		cr.Put("/ad", adHandler.UpdateADConfig)
		cr.Post("/ad/test", adHandler.TestADConfig)
	})

	// Permission administration is reserved for SUPERADMIN; permission:manage is not delegable.
	r.Route("/api/admin/permissions", func(pr chi.Router) {
		pr.With(authMw.Authenticate, auth.RequireRole(auth.RoleSuperadmin)).Get("/", permissionHandler.List)
	})
	r.With(authMw.Authenticate, auth.RequireRole(auth.RoleSuperadmin)).Put("/api/admin/roles/{role}/permissions", permissionHandler.UpdateRole)

	// User administration routes (Admin only).
	r.Route("/api/admin/users", func(ur chi.Router) {
		ur.Use(authMw.Authenticate)
		ur.Use(auth.RequirePermission(auth.PermissionUserManage))

		ur.Get("/", userAdminHandler.ListUsers)
		ur.Patch("/{id}", userAdminHandler.UpdateUser)
	})
	r.Route("/api/admin/locations/{code}/members", func(lr chi.Router) {
		lr.Use(authMw.Authenticate)
		lr.Use(auth.RequirePermission(auth.PermissionUserManage))
		lr.Get("/", locationMembershipHandler.ListLocationMembers)
		lr.Put("/{userID}", locationMembershipHandler.UpsertLocationMember)
		lr.Delete("/{userID}", locationMembershipHandler.DeleteLocationMember)
	})
	r.Route("/api/admin/teams/{id}/locations", func(tr chi.Router) {
		tr.Use(authMw.Authenticate)
		tr.Use(auth.RequirePermission(auth.PermissionUserManage))
		tr.Get("/", teamLocationHandler.ListTeamLocations)
		tr.Put("/{code}", teamLocationHandler.AddTeamLocation)
		tr.Delete("/{code}", teamLocationHandler.DeleteTeamLocation)
	})
	r.Route("/api/admin/settings", func(sr chi.Router) {
		sr.Use(authMw.Authenticate)
		sr.Use(auth.RequirePermission(auth.PermissionUserManage))
		sr.Get("/timezone", settingsHandler.GetTimezone)
		sr.Patch("/timezone", settingsHandler.UpdateTimezone)
	})

	registerBusinessRoutes(ctx, r, queries, authMw, cipher, cfg)
}

func registerBusinessRoutes(ctx context.Context, r *chi.Mux, queries *db.Queries, authMw *auth.Middleware, cipher *crypto.Cipher, cfg *config.Config) {
	storageDir := cfg.DataDir
	if storageDir == "" {
		storageDir = "./data"
	}
	storageMgr, err := storage.NewManager(storageDir)
	if err != nil {
		log.Printf("Warning: failed to init storage manager: %v", err)
	}

	registerMasterDataRoutes(r, queries, authMw)

	notifyCh := make(chan struct{}, 10)
	registerIssueRoutes(r, queries, storageMgr, authMw, notifyCh)
	registerScoringAndNotificationRoutes(ctx, r, queries, authMw, cipher, notifyCh, storageDir)
}

func registerMasterDataRoutes(r *chi.Mux, queries *db.Queries, authMw *auth.Middleware) {
	mdHandler := masterdata.NewHandler(queries)
	r.Route("/api/locations", func(lr chi.Router) {
		lr.Use(authMw.Authenticate)
		lr.Get("/", mdHandler.ListLocations)
		lr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Get("/all", mdHandler.ListAllLocations)
		lr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Post("/", mdHandler.CreateLocation)
		lr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Patch("/{code}/status", mdHandler.UpdateLocationStatus)
		lr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Put("/{code}", mdHandler.UpdateLocation)
	})
	r.Route("/api/tags", func(tr chi.Router) {
		tr.Use(authMw.Authenticate)
		tr.Get("/", mdHandler.ListTags)
		tr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Get("/all", mdHandler.ListAllTags)
		tr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Post("/", mdHandler.UpsertTag)
		tr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Patch("/{code}/status", mdHandler.UpdateTagStatus)
		tr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Patch("/{code}/review", mdHandler.ReviewTag)
		tr.With(auth.RequirePermission(auth.PermissionMasterdataManage)).Post("/batch-status", mdHandler.BatchUpdateTagsStatus)
	})
	respHandler := masterdata.NewResponsibilityHandler(queries)
	r.Route("/api/assets", func(ar chi.Router) {
		ar.Use(authMw.Authenticate)
		ar.Get("/", respHandler.ListAssets)
	})
	r.Route("/api/teams", func(tr chi.Router) {
		tr.Use(authMw.Authenticate)
		tr.Get("/", respHandler.ListTeams)
		tr.With(auth.RequirePermission(auth.PermissionIssueAssign)).Get("/{id}/members", respHandler.ListMembers)
	})
	r.Route("/api/admin/assets", func(ar chi.Router) {
		ar.Use(authMw.Authenticate)
		ar.Use(auth.RequirePermission(auth.PermissionMasterdataManage))
		ar.Post("/", respHandler.CreateAsset)
		ar.Put("/{id}", respHandler.UpdateAsset)
	})
	r.Route("/api/admin/teams", func(tr chi.Router) {
		tr.Use(authMw.Authenticate)
		tr.Use(auth.RequirePermission(auth.PermissionUserManage))
		tr.Post("/", respHandler.CreateTeam)
		tr.Put("/{id}", respHandler.UpdateTeam)
		tr.Get("/{id}/members", respHandler.ListMembers)
		tr.Put("/{id}/members/{userID}", respHandler.AddMember)
		tr.Delete("/{id}/members/{userID}", respHandler.DeleteMember)
	})
}

func registerIssueRoutes(r *chi.Mux, queries *db.Queries, storageMgr *storage.Manager, authMw *auth.Middleware, notifyCh chan struct{}) {
	if storageMgr == nil {
		return
	}
	issueSvc := issue.NewService(queries, storageMgr, notifyCh)
	issueHandler := issue.NewHandler(issueSvc)

	r.Route("/api/issues", func(ir chi.Router) {
		ir.Use(authMw.Authenticate)
		ir.Get("/events", issueHandler.Events)
		ir.Get("/", issueHandler.List)
		ir.Post("/sync", issueHandler.Sync)
		ir.Get("/{id}/media/{folder}/{filename}", issueHandler.Media)
		ir.Get("/{id}", issueHandler.GetByID)
		ir.Post("/{id}/resolve", issueHandler.Resolve)
		ir.Post("/{id}/close", issueHandler.Close)
		ir.Post("/{id}/reopen", issueHandler.Reopen)
		ir.Post("/{id}/invalidate", issueHandler.Invalid)
		ir.Patch("/{id}", issueHandler.Patch)
	})

	reportSvc := report.NewService(queries)
	reportHandler := report.NewHandler(reportSvc)

	r.Route("/api/reports", func(rr chi.Router) {
		rr.Use(authMw.Authenticate)
		rr.Use(auth.RequirePermission(auth.PermissionReportsView))
		rr.Get("/summary", reportHandler.GetSummary)
		rr.Get("/teams", reportHandler.GetTeams)
	})

	// Reports and exports require the reports capability.
	r.With(authMw.Authenticate, auth.RequirePermission(auth.PermissionReportsExport)).Get("/api/issues/export", reportHandler.ExportXLSX)
}

func registerScoringAndNotificationRoutes(ctx context.Context, r *chi.Mux, queries *db.Queries, authMw *auth.Middleware, cipher *crypto.Cipher, notifyCh chan struct{}, storageDir string) {
	scoringSvc := scoring.NewService(queries, nil)
	scoringHandler := scoring.NewHandler(scoringSvc)
	r.Route("/api/leaderboard", func(lbr chi.Router) {
		lbr.Use(authMw.Authenticate)
		lbr.Use(auth.RequirePermission(auth.PermissionReportsView))
		lbr.Get("/locations", scoringHandler.GetLocationLeaderboard)
		lbr.Get("/reporters", scoringHandler.GetReporterLeaderboard)
		lbr.Get("/score-logs", scoringHandler.GetTargetScoreLogs)
	})

	r.Route("/api/issues/{id}/score-logs", func(ilr chi.Router) {
		ilr.Use(authMw.Authenticate)
		ilr.Get("/", scoringHandler.GetIssueScoreLogs)
	})

	r.Route("/api/config/scoring", func(scr chi.Router) {
		scr.Use(authMw.Authenticate)
		scr.Get("/", scoringHandler.GetRules)
		scr.With(auth.RequirePermission(auth.PermissionScoringManage)).Put("/", scoringHandler.UpdateRules)
	})

	httpSender := notification.NewHTTPSender(cipher)
	notifHandler := notification.NewConfigHandler(queries, cipher, httpSender)
	r.Route("/api/config/notifications", func(nr chi.Router) {
		nr.Use(authMw.Authenticate)
		nr.Use(auth.RequireRole(auth.RoleAdmin))
		nr.Get("/", notifHandler.GetConfig)
		nr.Put("/", notifHandler.UpdateConfig)
		nr.Post("/test", notifHandler.TestConfig)
	})

	aiSvc := ai.NewService(queries, cipher, nil, storageDir)
	aiHandler := ai.NewHandler(aiSvc)
	r.Route("/api/config/ai", func(air chi.Router) {
		air.Use(authMw.Authenticate)
		air.Use(auth.RequireRole(auth.RoleAdmin, auth.RoleSuperadmin))
		air.Get("/", aiHandler.GetConfig)
		air.Put("/", aiHandler.UpdateConfig)
		air.Post("/test", aiHandler.TestConnection)
		air.Post("/test-dns", aiHandler.TestDNS)
	})

	r.Route("/api/ai", func(air chi.Router) {
		air.Use(authMw.Authenticate)
		air.Get("/status", aiHandler.Status)
		air.Post("/translate", aiHandler.Translate)
		air.Get("/cached", aiHandler.GetCached)
		air.Post("/review", aiHandler.Review)
		air.Post("/review-follow-up", aiHandler.FollowUp)
		air.Post("/suggest-tags", aiHandler.SuggestTags)
	})
	outboxWorker := notification.NewWorker(queries, httpSender, cipher, notifyCh)
	go outboxWorker.Start(ctx)

	factoryLoc := time.UTC
	if settings, err := queries.GetSystemSettings(ctx); err == nil {
		if loaded, loadErr := time.LoadLocation(settings.Timezone); loadErr == nil {
			factoryLoc = loaded
		} else {
			log.Printf("Warning: invalid factory timezone %q; using UTC: %v", settings.Timezone, loadErr)
		}
	}
	cronRunner := cron.NewRunner(queries, factoryLoc, storageDir)
	go cronRunner.Start(ctx)
}

func timeoutByRoute(next http.Handler) http.Handler {
	timeout := middleware.Timeout(60 * time.Second)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if r.URL.Path == "/api/issues/events" {
			next.ServeHTTP(w, r)
			return
		}
		timeout(next).ServeHTTP(w, r)
	})
}

func requestLogger(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		requestID := middleware.GetReqID(r.Context())
		if requestID != "" {
			w.Header().Set("X-Request-ID", requestID)
		}
		started := time.Now()
		ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
		next.ServeHTTP(ww, r)
		observability.Log("info", "http request", map[string]any{
			"request_id": requestID, "route": r.URL.Path, "status": ww.Status(),
			"duration_ms": time.Since(started).Seconds() * 1000,
		})
	})
}