# openapi.yaml
openapi: 3.1.0
info:
  title: 6S Issue Management & Visual Reporting System API
  version: 1.0.0
  description: API phục vụ ghi nhận, đồng bộ offline, giải quyết và tính điểm sức khỏe 6S trong môi trường nhà xưởng.
servers:
  - url: /api
    description: Runtime-configured API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    UserSummary:
      type: object
      required:
        - id
        - username
        - auth_source
        - role
        - capabilities
        - full_name
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: leader_a2
        auth_source:
          type: string
          enum: [LOCAL, AD]
          example: LOCAL
        role:
          type: string
          enum: [USER, LINE_LEADER, SAFETY_OFFICER, ADMIN, SUPERADMIN]
          example: LINE_LEADER
        capabilities:
          type: array
          items:
            type: string
          description: Effective resource:action capabilities. Missing capabilities deny UI access.
          example: [reports:view, issue:close_line]
        assigned_location_code:
          type: string
          nullable: true
          example: LINE_A2
        full_name:
          type: string
          example: Nguyễn Văn A
        email:
          type: string
          nullable: true
          example: nva@factory.lan
...
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      operationId: login
      summary: Đăng nhập tài khoản Local hoặc Active Directory / LDAP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: P@ssw0rd123
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - access_token
                      - expires_in
                      - refresh_token
                      - refresh_expires_in
                      - user
                    properties:
                      access_token:
                        type: string
                      expires_in:
                        type: integer
                        example: 3600
                      refresh_token:
                        type: string
                      refresh_expires_in:
                        type: integer
                        example: 2592000
                      user:
                        $ref: '#/components/schemas/UserSummary'
        '401':
          description: Sai thông tin đăng nhập
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Vượt quá giới hạn thử đăng nhập (Rate limit 5 lần / 1 phút)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/refresh:
    post:
      operationId: refresh
      summary: Cấp mới Access Token bằng Refresh Token (Token Rotation)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token được làm mới thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - access_token
                      - expires_in
                      - refresh_token
                      - refresh_expires_in
                      - user
                    properties:
                      access_token:
                        type: string
                      expires_in:
                        type: integer
                        example: 3600
                      refresh_token:
                        type: string
                      refresh_expires_in:
                        type: integer
                        example: 2592000
                      user:
                        $ref: '#/components/schemas/UserSummary'

        '401':
          description: Refresh token đã hết hạn hoặc bị thu hồi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/revoke:
    post:
      operationId: revokeSession
      summary: Thu hồi phiên đăng nhập (Đăng xuất hoặc thu hồi khẩn cấp)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token_id:
                  type: integer
                user_id:
                  type: integer
      responses:
        '200':
          description: Thu hồi thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      revoked:
                        type: boolean
                        example: true
  /auth/sessions:
    get:
      operationId: listAuthSessions
      summary: Liệt kê các phiên đăng nhập đang hoạt động
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
          description: Chỉ định user_id khi caller là Admin
      responses:
        '200':
          description: Danh sách phiên đăng nhập
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - created_at
                        - expires_at
                      properties:
                        id:
                          type: integer
                          example: 123
                        device_info:
                          type: string
                          nullable: true
                          example: Mobile Safari / iOS 17
                        created_at:
                          type: string
                          format: date-time
                        expires_at:
                          type: string
                          format: date-time
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /locations:
    get:
      operationId: listLocations
      summary: Lấy danh mục vị trí chuẩn hóa (Master Data)
      responses:
        '200':
          description: Danh sách vị trí
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
    post:
      operationId: createLocation
      summary: Thêm mới vị trí (Requires permission `masterdata:manage`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name_vi
                - name_zh
                - name_en
                - qr_code
              properties:
                code:
                  type: string
                  example: LINE_A3
                name_vi:
                  type: string
                  example: Chuyền May A3
                name_zh:
                  type: string
                  example: 缝纫 A3 线
                name_en:
                  type: string
                  example: Sewing Line A3
                qr_code:
                  type: string
                  example: LOC:LINE_A3
      responses:
        '201':
          description: Vị trí đã được tạo thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '403':
          description: Requires the route-specific management permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /tags:
    get:
      operationId: listTags
      summary: Lấy danh mục tags đa ngôn ngữ theo tần suất sử dụng
      responses:
        '200':
          description: Danh sách tags
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
    post:
      operationId: upsertTag
      summary: Thêm hoặc cập nhật tag (Requires permission `masterdata:manage`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name_vi
                - name_zh
                - name_en
                - category
              properties:
                code:
                  type: string
                  example: fire_hazard
                name_vi:
                  type: string
                  example: Nguy cơ cháy
                name_zh:
                  type: string
                  example: 火灾隐患
                name_en:
                  type: string
                  example: Fire hazard
                category:
                  type: string
                  enum: [1S, 2S, 3S, 4S, 5S, 6S]
                  example: 6S
      responses:
        '201':
          description: Tag đã được lưu thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
  /assets:
    get:
      operationId: listAssets
      summary: Danh mục thiết bị/tài sản theo vị trí (yêu cầu đăng nhập)
      parameters:
        - name: site_id
          in: query
          required: false
          description: Chỉ được khác site hiện tại với SUPERADMIN; nếu không sẽ trả 403
          schema:
            type: integer
        - name: location_code
          in: query
          required: false
          schema:
            type: string
        - name: is_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Danh sách thiết bị
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: site_id không thuộc phạm vi người dùng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /teams:
    get:
      operationId: listTeams
      summary: Danh mục team của site (yêu cầu đăng nhập)
      parameters:
        - name: site_id
          in: query
          required: false
          description: Chỉ được khác site hiện tại với SUPERADMIN; nếu không sẽ trả 403
          schema:
            type: integer
        - name: location_code
          in: query
          required: false
          description: Lọc team có gán khu vực này
          schema:
            type: string
        - name: is_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Danh sách team
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: site_id không thuộc phạm vi người dùng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /teams/{id}/members:
    get:
      operationId: listAssignableTeamMembers
      summary: Thành viên team phục vụ chọn người xử lý (Requires `issue:assign`)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách thành viên team
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamMember'
        '403':
          description: Thiếu permission `issue:assign`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Team không thuộc site của người dùng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /issues:
    get:
      operationId: listIssues
      summary: Lấy danh sách issues có phân trang và bộ lọc
      parameters:
        - name: status
          in: query
          description: Chấp nhận cả alias số nhiều `statuses` (giá trị lặp lại hoặc phân tách bằng dấu phẩy)
          schema:
            type: string
            enum: [OPEN, PENDING_REVIEW, CLOSED, INVALID]
        - name: category
          in: query
          description: Chấp nhận cả alias số nhiều `categories`
          schema:
            type: string
            enum: [1S, 2S, 3S, 4S, 5S, 6S]
        - name: location_code
          in: query
          description: Chấp nhận cả alias số nhiều `location_codes`
          schema:
            type: string
        - name: tag_code
          in: query
          required: false
          description: Lọc issue có gắn tag theo mã code
          schema:
            type: string
        - name: overdue
          in: query
          required: false
          description: Chỉ issue OPEN quá 48 giờ
          schema:
            type: boolean
        - name: assigned_team_id
          in: query
          required: false
          schema:
            type: integer
          description: Lọc theo team xử lý
        - name: mine_team
          in: query
          required: false
          schema:
            type: boolean
          description: Chỉ issues được giao cho một team hiện tại của người dùng
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Danh sách issues phân trang
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Issue'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /issues/{id}/resolve:
    post:
      operationId: resolveIssue
      summary: Upload ảnh khắc phục (hỗ trợ offline sync với expected_version)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - resolved_client_uuid
                - expected_version
                - photo_after
              properties:
                resolved_client_uuid:
                  type: string
                  format: uuid
                expected_version:
                  type: integer
                  example: 1
                photo_after:
                  type: string
                  format: binary
      responses:
        '200':
          description: Khắc phục thành công, chuyển sang PENDING_REVIEW
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Issue'
        '409':
          description: Xung đột trạng thái hoặc phiên bản (Optimistic Locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /admin/locations/{code}/members:
    get:
      operationId: listLocationMembers
      summary: Danh sách người phụ trách khu vực (Requires `user:manage`)
      parameters:
        - {name: code, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Danh sách thành viên}
...